feat: allow parallel requests by caching device info
This adds a simple in-memory cache to ensure we don't spawn a ton of `xbutil` processes if the number of requests increases. The default TTL is set to 5 seconds, with a command-line option `--xrt.cache-ttl` to tune it.

Hence, the maximum number of requests is now set to `0`, effectively disabling it to allow for maximum availability.